move set_children helper function to test.h
